requirements_pipecat.txt
# Pipecat WebRTC dependencies for MuseTalk
pipecat-ai>=0.0.40
daily-python>=0.10.0

# Audio processing
numpy>=1.21.0
librosa>=0.10.0
soundfile>=0.12.0

# Existing MuseTalk dependencies (if not already installed)
torch>=1.13.0
torchvision>=0.14.0
torchaudio>=0.13.0
transformers>=4.21.0
opencv-python>=4.6.0
omegaconf>=2.2.0
imageio>=2.21.0
tqdm>=4.64.0
Pillow>=9.2.0
scipy>=1.9.0
scikit-image>=0.19.0